+++ Koishibot.Core/Features/TwitchUsers/Models/TwitchUser.cs
@@ -5,6 +5,8 @@ using Koishibot.Core.Features.Common.Models;
 using Koishibot.Core.Features.Raids.Models;
 using Koishibot.Core.Features.RaidSuggestions.Models;
 using Koishibot.Core.Features.Supports.Models;
+using Koishibot.Core.Persistence;
+using Koishibot.Core.Persistence.Cache.Enums;
 using Koishibot.Core.Services.TwitchApi.Models;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
@@ -79,6 +81,69 @@ public class TwitchUser
 		return this;
 	}
 }
+
+public static class TwitchUserExtension
+{
+	// CACHE
+
+	public static TwitchUser? FindUserByTwitchId(this IAppCache cache, string id)
+	{
+		var userList = cache.Get<List<TwitchUser>>(CacheName.Users)
+		               ?? new List<TwitchUser>();
+
+		// userList ??= new List<TwitchUser>();
+		return userList.Find(x => x.TwitchId == id);
+	}
+
+	public static void AddUser(this IAppCache cache, TwitchUser user)
+	{
+		var userList = cache.Get<List<TwitchUser>>(CacheName.Users)
+		               ?? new List<TwitchUser>();
+
+		userList.Add(user);
+		cache.Add(CacheName.Users, userList, TimeSpan.FromDays(1));
+	}
+
+	public static void UpdateUser(this IAppCache cache, TwitchUser user)
+	{
+		var userList = cache.Get<List<TwitchUser>>(CacheName.Users)
+		               ?? new List<TwitchUser>();
+
+		var index = userList!.FindIndex(x => x.TwitchId == user.TwitchId);
+
+		if (index != -1)
+		{
+			userList[index] = user;
+		}
+
+		cache.Add(CacheName.Users, cache, TimeSpan.FromDays(1));
+	}
+
+
+	public static async Task<TwitchUser?> GetUserByTwitchId
+	(this KoishibotDbContext Context, string twitchId)
+	{
+		return await Context.Users
+		.FirstOrDefaultAsync(tu => tu.TwitchId == twitchId);
+	}
+
+	public static async Task<TwitchUser?> GetUserByLogin
+	(this KoishibotDbContext context, string userlogin)
+	{
+		return await context.Users
+		.FirstOrDefaultAsync(x => x.Login == userlogin);
+	}
+
+	public static async Task<TwitchUser> UpdateUser
+	(this KoishibotDbContext database, TwitchUser user)
+	{
+		database.Update(user);
+		await database.SaveChangesAsync();
+		return user;
+	}
+}
+
+
 
 /*══════════════════【 CONFIGURATION 】═════════════════*/
 public class UserConfig : IEntityTypeConfiguration<TwitchUser>
