+++ Koishibot.Core/Features/TwitchUsers/TwitchUserRegistration.cs
@@ -1,5 +1,4 @@
 ﻿using Koishibot.Core.Features.ChatCommands.Models;
-using Koishibot.Core.Features.TwitchUsers.Extensions;
 using Koishibot.Core.Features.TwitchUsers.Interfaces;
 using Koishibot.Core.Features.TwitchUsers.Models;
 using Koishibot.Core.Persistence;
